refactor(perf): split /ai-architectures + /developers into server shell + motion island
15th + 16th in the pattern series.
