+++ components/ai-architectures/architectures-data.ts
@@ -1,0 +1,32 @@
+export const architecturesListItemData = [
+  {
+    id: 'enterprise-rag',
+    title: 'Enterprise RAG Platform',
+    prototypeSlug: 'enterprise-rag-platform',
+  },
+  {
+    id: 'multi-agent-orchestrator',
+    title: 'Multi-Agent Code Assistant',
+    prototypeSlug: 'multi-agent-code-assistant',
+  },
+  {
+    id: 'ai-gateway',
+    title: 'Enterprise AI Gateway',
+    prototypeSlug: null,
+  },
+  {
+    id: 'llmops-pipeline',
+    title: 'LLMOps Pipeline',
+    prototypeSlug: null,
+  },
+  {
+    id: 'mcp-hub',
+    title: 'MCP Server Hub',
+    prototypeSlug: null,
+  },
+  {
+    id: 'ai-coe',
+    title: 'AI Center of Excellence',
+    prototypeSlug: null,
+  },
+]
