+++ components/developers/developers-data.ts
@@ -1,0 +1,20 @@
+export const developersLearningPaths = [
+  {
+    title: 'Claude Code Mastery',
+    description:
+      'From basic prompting to autonomous agent workflows. Learn the patterns that 10x your coding velocity.',
+    href: '/prompt-library',
+  },
+  {
+    title: 'Agentic Patterns',
+    description:
+      'Build production-grade autonomous systems. MCP servers, tool chains, and multi-agent coordination.',
+    href: '/prompt-library',
+  },
+  {
+    title: 'Workflow Automation',
+    description:
+      'Automate your entire development lifecycle. From PR reviews to deployment pipelines.',
+    href: '/prompt-library',
+  },
+]
